+++ EL_CfitsioWrapper/tests/src/HduWrapper_test.cpp
@@ -53,17 +53,6 @@ BOOST_FIXTURE_TEST_CASE(create_and_access_image_extension_test, FitsIO::Test::Mi
   BOOST_TEST(output.vector() == input.vector());
 }
 
-BOOST_AUTO_TEST_CASE(category_ordering_test) {
-  using namespace FitsIO;
-  BOOST_TEST(isInstance<HduCategory::Primary>(HduCategory::MetadataPrimary));
-  BOOST_TEST(not isInstance<HduCategory::MetadataPrimary>(HduCategory::Primary));
-  BOOST_TEST(not isInstance<HduCategory::Primary>(HduCategory::Image));
-  BOOST_TEST(isInstance<HduCategory::Any>(HduCategory::Image));
-  BOOST_TEST(isInstance<HduCategory::Any>(HduCategory::Bintable));
-  BOOST_TEST(isInstance<HduCategory::Image>(HduCategory::FloatImage));
-  BOOST_TEST(isInstance<HduCategory::ImageExt>(HduCategory::FloatImage | HduCategory::Ext));
-}
-
 //-----------------------------------------------------------------------------
 
 BOOST_AUTO_TEST_SUITE_END()